backend/pyproject.toml
[project]
name = "ocrcheck-backend"
version = "0.1.0"
description = "OCRCheck - Document scanning, OCR, and management system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "boto3>=1.35.0",
    "aioboto3>=13.0.0",
    "redis>=5.2.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "rapidocr_onnxruntime>=1.4.0",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "anthropic>=0.42.0",
    "opensearch-py>=2.7.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
